+++ package.json
@@ -43,6 +43,7 @@
     "@uniswap/universal-router-sdk": "^2.2.0",
     "@uniswap/v2-sdk": "^4.3.2",
     "@uniswap/v3-sdk": "^3.13.0",
+    "@uniswap/v4-sdk": "^1.0.0",
     "async-retry": "^1.3.1",
     "await-timeout": "^1.1.1",
     "axios": "^0.21.1",
